// generated/api/apiSchemas.ts
export type EntityTypedMetadata = Record<string, any> & CommonMetadata;

/**
 * Metadata Claims that can be used with any Entity Type
 */
export type CommonMetadata = {
  /**
   * A human-readable name representing the organization owning this Entity. If the owner is a physical person, this MAY be, for example, the person's name.
   */
  organization_name?: string;
  /**
   * A human-readable name of the Entity to be presented to the End-User.
   */
  display_name?: string;
  /**
   * A human-readable brief description of this Entity presentable to the End-User.
   */
  description?: string;
  /**
   * JSON array with one or more strings representing search keywords, tags, categories, or labels that apply to this Entity.
   */
  keywords?: string[];
  /**
   * JSON array with one or more strings representing contact persons at the Entity. These MAY contain names, e-mail addresses, descriptions, phone numbers, etc.
   */
  contacts?: string[];
  /**
   * A URL that points to the logo of this Entity. The file containing the logo SHOULD be published in a format that can be viewed via the web.
   *
   * @format uri
   */
  logo_uri?: string;
  /**
   * URL of the documentation of conditions and policies relevant to this Entity.
   *
   * @format uri
   */
  policy_uri?: string;
  /**
   * URL for documentation of additional information about this Entity viewable by the End-User.
   *
   * @format uri
   */
  information_uri?: string;
  /**
   * URL of a Web page for the organization owning this Entity.
   *
   * @format uri
   */
  organization_uri?: string;
};

/**
 * JSON object that declares roles that the Entity plays - its Entity Types - and that contains metadata for those Entity Types. Each member name of the JSON object is an Entity Type Identifier, and each value MUST be a JSON object containing metadata parameters according to the metadata schema of the Entity Type.
 */
export type Metadata = {
  [key: string]: EntityTypedMetadata;
};